update post-install script again

+++ meta-container-os/dynamic-layers/raspberrypi/recipes-core/bundles/files/post-install.sh
@@ -8,6 +8,12 @@ safe_copy() {
     echo "safe_copy can only handle one file copy at a time" >&2
     exit 2
   fi
+
+  # do not copy if the files are the same
+  if cmp -s "$1" "$2"; then
+    return
+  fi
+
   cp -a "$1" "$2".tmp || return $?
   sync "$2".tmp || return $?
   mv "$2".tmp "$2" || return $?
@@ -17,9 +23,6 @@ safe_copy() {
 update_firmware() {
   RAUC_MOUNT_POINT="$1"
   echo "[RAUC Hook] Updating firmware files from $RAUC_MOUNT_POINT/boot.image to /boot"
-
-  #echo "$RAUC_MOUNT_POINT" > "$RAUC_MOUNT_POINT/root/.rauc-mount-point"
-  #echo "$RAUC_MOUNT_POINT" >> "/data/rauc-mount-point"
 
   # Copy 'core' firmware files first
   find "$RAUC_MOUNT_POINT/boot.image" -maxdepth 1 -type f ! -name "*.dtbo" | while IFS= read -r f; do
